Multiple NACV derivative integral evaluations merged into one kernel to avoid recomputation (#631)
* Optimize the DF-TDDFT gradients and NAC by processing the derivatives of multiple density matrices in one kernel

* Fix int3c2e_contract_dm

* Bugfix

* fix tests

* Optimize contract_int3c2e_dm

* Miss factor 1/2 in tdrhf_nac and tddft_ris_nac

* Bugfix for tdrks_ris_grad

* Update tests

+++ gpu4pyscf/lib/gvhf-rys/ejk_int3c2e_ip1.cu
@@ -173,7 +173,7 @@ void ejk_int3c2e_ip1_kernel(double *ejk, double *ejk_aux,
                         int i0 = envs.ao_loc[ish];
                         int j0 = envs.ao_loc[jsh];
                         int k0 = envs.ao_loc[ksh] - nao;
-                        double *dm_local = dm + (i_dm * nao + j0) * nao + i0;
+                        double *dm_local = dm + (i_dm * nao + j0) * (size_t)nao + i0;
 #pragma unroll
                         for (int n = 0; n < GOUT_WIDTH; ++n) {
                             uint32_t ijk = n*gout_stride+gout_id;
